pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "makelive"
authors = [{ name = "Rhet Turnbull", email = "rturnbull+git@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">3.9"
dependencies = [
    "pyobjc-core>=9.2",
    "pyobjc-framework-AVFoundation>=9.2",
    "pyobjc-framework-Contacts>=9.2",
    "pyobjc-framework-CoreLocation>=9.2",
    "pyobjc-framework-Quartz>=9.2",
    "pyobjc-framework-UniformTypeIdentifiers>=9.2",
    "wheel>=0.41.2",
    "wurlitzer>=3.0.3",
]

[project.optional-dependencies]
test = ["pytest>=7.4.2", "pytest-cov", "mypy>=1.6.1"]
lint = ["ruff>=0.1.14"]

# docs = [
#     "mkdocs>=1.4.2",
#     "mkdocs-material>=9.0.13",
#     "mkdocstrings-python>=0.8.3",
# ]

[project.urls]
Home = "https://github.com/RhetTbull/makelive"
Issues = "https://github.com/RhetTbull/makelive/issues"
Source = "https://github.com/RhetTbull/makelive"


[project.scripts]
makelive = "makelive:main"

[tool.flit.sdist]
exclude = ["site/**/*", "tests/**/*"]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 130
